+++ dependency_injector/catalogs/__init__.py
@@ -1,6 +1,5 @@
 """Dependency injector catalogs package."""
 
-from dependency_injector.catalogs.bundle import CatalogBundle
 from dependency_injector.catalogs.dynamic import DynamicCatalog
 from dependency_injector.catalogs.declarative import (
     DeclarativeCatalogMetaClass,
@@ -13,7 +12,6 @@ from dependency_injector.catalogs.utils import (
 
 
 __all__ = (
-    'CatalogBundle',
     'DynamicCatalog',
     'DeclarativeCatalogMetaClass',
     'DeclarativeCatalog',